refactor(settings): decompose Settings into focused sub-components
Closes #215.

Settings.tsx shrinks from 368 → 55 lines by splitting into:
- GitHubSettings: auth + poll interval
- WorkspaceSettings: max active workspaces
- RepositorySettings: repo list, filter, batch actions
- NumberField: shared numeric input
- useConfigMutation: shared config mutation hook
- styles: shared section class

NumberField now resets via React `key` prop instead of a useEffect-driven
counter, aligning with the React mental model. Each section owns its own
reset signal locally, and the parent only holds the `saveError` state and
passes a single `onError` callback to each child. Test reference to the
poll-interval input is re-queried after remount.
