+++ src/components/Settings/styles.ts
@@ -1,0 +1,1 @@
+export const sectionClass = "flex flex-col gap-3 border-b border-border pb-4";
